fixed some bugs

+++ core/src/splinetlsm/coefs.cpp
@@ -60,7 +60,7 @@ namespace splinetlsm {
         }
         
         // re-weight sample for sampling of time points
-        double time_weight = Y.n_elem / sample_info.time_indices.n_elem;
+        double time_weight = (double) Y.n_elem / sample_info.time_indices.n_elem;
         grad_mean *= 0.5 * time_weight;
         grad_prec *= 0.5 * time_weight;
